// src/model/PackageModel.ts
export interface PackageModel {
    id?: number;
    packageName?: string;
    system?: string;
    carValue?: string;
    userValue?: string;
    mobileSupport?: string;
    serverType?: string;
    backup?: string;
    support?: string;
    shipmentValue?: string;
    cargoList?: string;
    promotion?: string;
    functionForConsumer?: string;
    viewMode?: boolean;
}